chore: clean up old schedule for new package
