+++ ui/components/schedule/client.go
