+++ ui/components/schedule/keymap.go
@@ -1,4 +1,4 @@
-package schedulelist
+package schedule
 
 import "charm.land/bubbles/v2/key"
