+++ ui/components/schedule/style.go
@@ -1,20 +1,7 @@
 package schedule
 
-import (
-	"os"
-
-	"charm.land/lipgloss/v2"
-	"github.com/KevinStirling/scorebug.sh/ui/components/scorebug"
-	"github.com/KevinStirling/scorebug.sh/ui/components/theme"
-)
+import "charm.land/lipgloss/v2"
 
 var (
-	grey          = lipgloss.Lighten(lipgloss.BrightBlack, .1)
-	divider       = lipgloss.NewStyle().Width(scorebug.SB_WIDTH + theme.Margin)
-	secondaryText = lipgloss.NewStyle().Foreground(grey)
-
-	hasDark          = lipgloss.HasDarkBackground(os.Stdin, os.Stdout)
-	lightDark        = lipgloss.LightDark(hasDark)
-	adaptiveInactive = lightDark(lipgloss.Black, lipgloss.BrightBlack)
-	adaptiveActive   = lightDark(grey, lipgloss.BrightWhite)
+	listStyle = lipgloss.NewStyle().Margin(0, 1, 0, 0)
 )
