--- ui/components/schedulelist/style.go
@@ -1,7 +1,0 @@
-package schedulelist
-
-import "charm.land/lipgloss/v2"
-
-var (
-	listStyle = lipgloss.NewStyle().Margin(0, 1, 0, 0)
-)
